add anthropic text provider
